requirements/base.txt
django==4.1.7
django-environ==0.10.0
psycopg2-binary==2.9.5
bs4==0.0.1
redis==4.4.2
Pillow==9.5.0
crispy-bootstrap4==2022.1
django-datatables-too==0.1.4
django-phonenumber-field==7.1.0
phonenumbers==8.13.13
django-allauth==0.52.0
django-extra-views==0.14.0
django-widget-tweaks==1.4.12
django-anymail[sendgrid]==9.1
geopy==2.3.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
python-barcode==0.14.0
django-cors-headers==4.2.0
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
num2words==0.5.13
reportlab==4.2.2
xhtml2pdf==0.2.16
googlemaps==4.10.0